requirements.txt
docling
google-genai
python-dotenv
psycopg2-binary